+++ src/renderer.js
@@ -1,24 +1,12 @@
-import "./components/section.js";
-import "./components/panel.js";
-import "./components/list.js";
-import "./components/item.js";
-import "./components/select.js";
-import "./components/option.js";
-import "./components/switch.js";
-import "./components/button.js";
-import "./components/text.js";
-import "./components/divider.js";
-
-import { SettingInterface } from "./setting/index.js";
+import { SettingInterface } from "./setting/renderer.js";
 
 
-const loader = new class {
-
+const loader = await (new class {
 
     #exports = {};
 
-
     async init() {
+        // 加载插件
         for (const [slug, plugin] of Object.entries(LiteLoader.plugins)) {
             if (plugin.disabled || plugin.incompatible) {
                 continue;
@@ -27,31 +15,24 @@ const loader = new class {
                 this.#exports[slug] = await import(`local:///${plugin.path.injects.renderer}`);
             }
         }
+        return this;
     }
 
-
     onSettingWindowCreated(settingInterface) {
-        for (const [slug, plugin] of Object.entries(LiteLoader.plugins)) {
-            if (plugin.disabled || plugin.incompatible) {
-                continue;
-            }
-            if (this.#exports?.[slug]?.onSettingWindowCreated) {
+        for (const [slug, plugin] of Object.entries(this.#exports)) {
+            if (plugin?.onSettingWindowCreated) {
                 const view = settingInterface.getSettingView(slug);
-                settingInterface.addNavItem(plugin.manifest.name, view);
-                this.#exports?.[slug]?.onSettingWindowCreated?.(view);
+                settingInterface.addNavItem(LiteLoader.plugins[slug].manifest.name, view);
+                plugin.onSettingWindowCreated(view);
             }
         }
     }
 
-
-};
-
-
-await loader.init();
+}).init();
 
 
-// 监听页面变化
-if (location.href.includes("/index.html") && location.href.includes("#/setting")) {
+// 注入设置界面
+function settingInterfaceInit() {
     const settingInterface = new SettingInterface();
     if (document.querySelector(".setting-tab .nav-bar")) {
         loader.onSettingWindowCreated(settingInterface);
@@ -69,18 +50,18 @@ if (location.href.includes("/index.html") && location.href.includes("#/setting")
         });
     }
 }
+
+
+// 监听页面变化
+if (location.href.includes("/index.html") && location.href.includes("#/setting")) {
+    settingInterfaceInit();
+}
 else {
     navigation.addEventListener("navigatesuccess", function func(event) {
         const url = event.target.currentEntry.url;
         if (url.includes("/index.html") && url.includes("#/setting")) {
             navigation.removeEventListener("navigatesuccess", func);
-            const settingInterface = new SettingInterface();
-            const interval = setInterval(() => {
-                if (document.querySelector(".setting-tab .nav-bar")) {
-                    clearInterval(interval);
-                    loader.onSettingWindowCreated(settingInterface);
-                }
-            }, 100);
+            settingInterfaceInit();
         }
     });
 }
