+++ src/setting/components/divider.js
