+++ src/setting/components/panel.js
