+++ src/setting/components/switch.js
