+++ src/setting/components/text.js
