+++ src/setting/renderer.js
@@ -1,3 +1,118 @@
+import "./components/section.js";
+import "./components/panel.js";
+import "./components/list.js";
+import "./components/item.js";
+import "./components/select.js";
+import "./components/option.js";
+import "./components/switch.js";
+import "./components/button.js";
+import "./components/text.js";
+import "./components/divider.js";
+
+
+export class SettingInterface {
+    #setting_title = document.querySelector(".setting-main .setting-title");
+    #liteloader_nav_bar = document.createElement("div");
+    #liteloader_setting_view = document.createElement("div");
+
+
+    constructor() {
+        const style = document.createElement("style");
+        style.textContent = `
+        .setting-tab {
+            display: flex;
+            flex-direction: column;
+        }
+        .setting-tab .nav-bar {
+            flex-shrink: 0;
+        }
+        .liteloader.nav-bar {
+            flex: 1;
+            margin-top: 25px;
+            overflow-x: hidden;
+            overflow-y: scroll;
+        }
+        .liteloader.nav-bar::before {
+            content: "";
+            display: block;
+            position: absolute;
+            transform: translate(50px, -15px);
+            width: 100px;
+            height: 5px;
+            border-radius: 5px;
+            background: rgba(127, 127, 127, 0.5);
+        }
+        .liteloader.tab-view {
+            font-size: 14px;
+            padding: 20px 20px 0px;
+        }
+        .liteloader.disabled {
+            pointer-events: none;
+            opacity: 0.5;
+        }
+        `;
+        document.head.append(style);
+
+        this.#liteloader_nav_bar.classList.add("nav-bar", "liteloader");
+        document.querySelector(".setting-tab").append(this.#liteloader_nav_bar);
+
+        this.#liteloader_setting_view.classList.add("q-scroll-view", "scroll-view--show-scrollbar");
+        document.querySelector(".setting-main .setting-main__content").append(this.#liteloader_setting_view);
+
+        const qqnt_setting_view = document.querySelector(".setting-main .q-scroll-view");
+        document.querySelector(".setting-tab").addEventListener("click", event => {
+            const nav_item = event.target.closest(".nav-item");
+            if (nav_item) {
+                // 重新设定激活状态
+                document.querySelectorAll(".setting-tab .nav-item").forEach(element => {
+                    element.classList.remove("nav-item-active");
+                });
+                nav_item.classList.add("nav-item-active");
+                // 内容显示
+                if (nav_item.classList.contains("liteloader")) {
+                    qqnt_setting_view.style.display = "none";
+                    this.#liteloader_setting_view.style.display = "block";
+                }
+                else {
+                    qqnt_setting_view.style.display = "block";
+                    this.#liteloader_setting_view.style.display = "none";
+                }
+            }
+        });
+
+        const view = this.getSettingView("config_view");
+        this.addNavItem("LiteLoaderQQNT", view);
+        onSettingWindowCreated(view);
+    }
+
+
+    // Setting View
+    getSettingView(slug) {
+        const view = document.createElement("div");
+        view.classList.add("liteloader", "tab-view", slug);
+        return view;
+    }
+
+
+    // 导航栏条目
+    addNavItem(name, view) {
+        const nav_item = document.querySelector(".setting-tab .nav-item").cloneNode(true);
+        nav_item.classList.remove("nav-item-active");
+        nav_item.classList.add("liteloader");
+        nav_item.querySelector(".q-icon").textContent = null;
+        nav_item.querySelector(".name").textContent = name;
+        nav_item.addEventListener("click", event => {
+            if (!event.currentTarget.classList.contains("nav-item-active")) {
+                this.#setting_title.childNodes[1].textContent = name;
+                this.#liteloader_setting_view.textContent = null;
+                this.#liteloader_setting_view.append(view);
+            }
+        });
+        this.#liteloader_nav_bar.append(nav_item);
+    }
+}
+
+
 // 对比本地与远端的版本号，有新版就返回true
 function compareVersion(local_version, remote_version) {
     // 将字符串改为数组
@@ -183,7 +298,7 @@ function initAbout(view) {
 }
 
 
-export async function onSettingWindowCreated(view) {
+async function onSettingWindowCreated(view) {
     // HTMl
     view.innerHTML = await (await fetch(`local:///${LiteLoader.path.root}/src/setting/static/view.html`)).text();
